# docs/FE/openapi.yaml
openapi: 3.0.3
info:
  title: Golf Store API Gateway
  version: 1.0.0
  description: API contract for FE integration via single API Gateway.
servers:
  - url: http://localhost:8080/api/v1
    description: API Gateway v1
security:
  - bearerAuth: []
...
paths:
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '423':
          $ref: '#/components/responses/AccountLockedError'
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Refresh success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshTokenResponseEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/logout:
    post:
      tags: [Auth]
      summary: Logout
      responses:
        '204':
          description: Logout success
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me:
    get:
      tags: [Profile]
      summary: Current user profile
      responses:
        '200':
          description: Profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfileEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /orders:
    post:
      tags: [Orders]
      summary: Create order
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrderRequest'
      responses:
        '201':
          description: Order created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummaryEnvelope'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '409':
          $ref: '#/components/responses/OutOfStockError'
    get:
      tags: [Orders]
      summary: List my orders
      parameters:
        - $ref: '#/components/parameters/OrderStatus'
        - $ref: '#/components/parameters/From'
        - $ref: '#/components/parameters/To'
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Order list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderListEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /orders/{order_id}:
    get:
      tags: [Orders]
      summary: Order detail
      parameters:
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: Order detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderDetailEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /orders/{order_id}/cancel:
    post:
      tags: [Orders]
      summary: Cancel order
      parameters:
        - $ref: '#/components/parameters/OrderIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelOrderRequest'
      responses:
        '200':
          description: Order canceled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderSummaryEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/OrderCannotCancelError'
  /orders/{order_id}/tracking:
    get:
      tags: [Orders]
      summary: Order tracking timeline
      parameters:
        - $ref: '#/components/parameters/OrderIdPath'
      responses:
        '200':
          description: Tracking data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OrderTrackingEnvelope'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    OrderIdPath:
      name: order_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
...
    Page:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    OrderStatus:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/OrderStatus'
...
    From:
      name: from
      in: query
      required: false
      schema:
        type: string
        format: date-time
    To:
      name: to
      in: query
      required: false
      schema:
        type: string
        format: date-time
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    AccountLockedError:
      description: Account locked
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    OutOfStockError:
      description: Product out of stock
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    OrderCannotCancelError:
      description: Order cannot be canceled
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    EnvelopeBase:
      type: object
      required: [success, request_id, timestamp]
      properties:
        success:
          type: boolean
          example: true
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
        meta:
          type: object
          additionalProperties: true
    ErrorEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            success:
              type: boolean
              enum: [false]
            error:
              $ref: '#/components/schemas/ErrorObject'
          required: [error]
    ErrorObject:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: RESOURCE_NOT_FOUND
        message:
          type: string
        details:
          type: array
          items:
            type: object
            additionalProperties: true
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
          format: password
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
    UserRole:
      type: string
      enum: [USER, ADMIN]
    UserStatus:
      type: string
      enum: [ACTIVE, LOCKED, DISABLED]
...
    OrderStatus:
      type: string
      enum: [NEW, PENDING_PAYMENT, PAID, PROCESSING, SHIPPING, COMPLETED, CANCELLED]
    PaymentStatus:
      type: string
      enum: [UNPAID, PAID, FAILED, REFUNDED]
    PaymentTxnState:
      type: string
      enum: [PENDING, SUCCESS, FAILED, CANCELLED]
...
    UserProfile:
      type: object
      required: [id, role, full_name, email, phone, status]
      properties:
        id:
          type: string
          format: uuid
        role:
          $ref: '#/components/schemas/UserRole'
        full_name:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          $ref: '#/components/schemas/UserStatus'
    AuthTokenData:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user]
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 3600
        user:
          $ref: '#/components/schemas/UserProfile'
    LoginResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/AuthTokenData'
    RefreshTokenResponseEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          required: [data]
          properties:
            data:
              type: object
              required: [access_token, refresh_token, token_type, expires_in]
              properties:
                access_token:
                  type: string
                refresh_token:
                  type: string
                token_type:
                  type: string
                expires_in:
                  type: integer
    UserProfileEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          required: [data]
          properties:
            data:
              $ref: '#/components/schemas/UserProfile'
...
    Pagination:
      type: object
      required: [page, page_size, total, total_pages]
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    ShippingAddressInput:
      type: object
      required: [recipient_name, recipient_phone, line1, city, country_code]
      properties:
        recipient_name:
          type: string
        recipient_phone:
          type: string
        line1:
          type: string
        line2:
          type: string
          nullable: true
        ward:
          type: string
          nullable: true
        district:
          type: string
          nullable: true
        city:
          type: string
        province:
          type: string
          nullable: true
        postal_code:
          type: string
          nullable: true
        country_code:
          type: string
    OrderItemInput:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
          format: uuid
        quantity:
          type: integer
          minimum: 1
    CreateOrderRequest:
      type: object
      required: [items, shipping_address]
      properties:
        items:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/OrderItemInput'
        shipping_address:
          $ref: '#/components/schemas/ShippingAddressInput'
        customer_note:
          type: string
          nullable: true
    OrderSummary:
      type: object
      required:
        [id, order_code, order_status, payment_status, subtotal_amount, discount_amount, shipping_fee, total_amount, placed_at]
      properties:
        id:
          type: string
          format: uuid
        order_code:
          type: string
        order_status:
          $ref: '#/components/schemas/OrderStatus'
        payment_status:
          $ref: '#/components/schemas/PaymentStatus'
        subtotal_amount:
          type: string
        discount_amount:
          type: string
        shipping_fee:
          type: string
        total_amount:
          type: string
        payment_due_at:
          type: string
          format: date-time
          nullable: true
        placed_at:
          type: string
          format: date-time
    OrderItem:
      type: object
      required: [product_id, sku, name, unit_price, quantity, line_total]
      properties:
        product_id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        unit_price:
          type: string
        quantity:
          type: integer
        line_total:
          type: string
    PaymentTransaction:
      type: object
      required: [id, status, txn_type, provider, amount, currency_code, created_at]
      properties:
        id:
          type: string
          format: uuid
        txn_type:
          type: string
          enum: [PAYMENT, REFUND]
        provider:
          type: string
        provider_txn_code:
          type: string
          nullable: true
        status:
          $ref: '#/components/schemas/PaymentTxnState'
        amount:
          type: string
        currency_code:
          type: string
        created_at:
          type: string
          format: date-time
    OrderDetail:
      allOf:
        - $ref: '#/components/schemas/OrderSummary'
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: '#/components/schemas/OrderItem'
            payments:
              type: array
              items:
                $ref: '#/components/schemas/PaymentTransaction'
    OrderSummaryEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderSummary'
    OrderListEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              type: object
              properties:
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/OrderSummary'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    OrderDetailEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderDetail'
    CancelOrderRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
    TrackingEvent:
      type: object
      required: [status, source_type, occurred_at]
      properties:
        status:
          type: string
        source_type:
          type: string
          enum: [SYSTEM, USER, ADMIN, PAYMENT_GATEWAY, CARRIER]
        occurred_at:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
    OrderTracking:
      type: object
      required: [order_id, current_status, timeline]
      properties:
        order_id:
          type: string
          format: uuid
        current_status:
          $ref: '#/components/schemas/OrderStatus'
        timeline:
          type: array
          items:
            $ref: '#/components/schemas/TrackingEvent'
    OrderTrackingEnvelope:
      allOf:
        - $ref: '#/components/schemas/EnvelopeBase'
        - type: object
          properties:
            data:
              $ref: '#/components/schemas/OrderTracking'
...